tests, warehouse: Restructure utilities
