+++ tests/unit/malware/test_init.py
@@ -170,5 +170,5 @@ def test_includeme(monkeypatch):
     malware.includeme(config)
 
     assert config.register_service_factory.calls == [
-        pretend.call(malware_check_class.create_service, IMalwareCheckService),
+        pretend.call(malware_check_class.create_service, IMalwareCheckService)
     ]
