--- tests/unit/malware/test_setup_patterns.py
@@ -1,243 +1,0 @@
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import io
-import tarfile
-import zipfile
-
-import pretend
-import pytest
-import yara
-
-from warehouse.malware.checks.setup_patterns import check as c
-from warehouse.malware.models import (
-    MalwareCheckState,
-    VerdictClassification,
-    VerdictConfidence,
-)
-
-from ...common.db.malware import MalwareCheckFactory
-from ...common.db.packaging import FileFactory
-
-
-def test_initializes(db_session):
-    check_model = MalwareCheckFactory.create(
-        name="SetupPatternCheck", state=MalwareCheckState.Enabled
-    )
-    check = c.SetupPatternCheck(db_session)
-
-    assert check.id == check_model.id
-    assert isinstance(check._yara_rules, yara.Rules)
-
-
-def test_extract_setup_contents_zip(db_session):
-    zipbuf = io.BytesIO()
-    with zipfile.ZipFile(zipbuf, mode="w") as zipobj:
-        zipobj.writestr("foo/setup.py", b"these are some contents")
-    zipbuf.seek(0)
-
-    MalwareCheckFactory.create(
-        name="SetupPatternCheck", state=MalwareCheckState.Enabled
-    )
-    check = c.SetupPatternCheck(db_session)
-
-    assert check._extract_setup_contents(zipbuf) == b"these are some contents"
-
-
-def test_extract_setup_contents_zip_no_setup(db_session):
-    zipbuf = io.BytesIO()
-    with zipfile.ZipFile(zipbuf, mode="w") as zipobj:
-        zipobj.writestr("foo/notsetup.py", b"these are some contents")
-    zipbuf.seek(0)
-
-    MalwareCheckFactory.create(
-        name="SetupPatternCheck", state=MalwareCheckState.Enabled
-    )
-    check = c.SetupPatternCheck(db_session)
-
-    assert check._extract_setup_contents(zipbuf) is None
-
-
-def test_extract_setup_contents_tar(db_session):
-    tarbuf = io.BytesIO()
-    with tarfile.open(fileobj=tarbuf, mode="w:gz") as tarobj:
-        contents = io.BytesIO(b"these are some contents")
-        member = tarfile.TarInfo(name="foo/setup.py")
-        member.size = len(contents.getbuffer())
-        tarobj.addfile(member, fileobj=contents)
-    tarbuf.seek(0)
-
-    MalwareCheckFactory.create(
-        name="SetupPatternCheck", state=MalwareCheckState.Enabled
-    )
-    check = c.SetupPatternCheck(db_session)
-
-    assert check._extract_setup_contents(tarbuf) == b"these are some contents"
-
-
-def test_extract_setup_contents_tar_empty(db_session):
-    tarbuf = io.BytesIO(b"invalid tar contents")
-    MalwareCheckFactory.create(
-        name="SetupPatternCheck", state=MalwareCheckState.Enabled
-    )
-    check = c.SetupPatternCheck(db_session)
-
-    assert check._extract_setup_contents(tarbuf) is None
-
-
-def test_extract_setup_contents_tar_no_setup(db_session):
-    tarbuf = io.BytesIO()
-    with tarfile.open(fileobj=tarbuf, mode="w:gz") as tarobj:
-        contents = io.BytesIO(b"these are some contents")
-        member = tarfile.TarInfo(name="foo/notsetup.py")
-        member.size = len(contents.getbuffer())
-        tarobj.addfile(member, fileobj=contents)
-    tarbuf.seek(0)
-
-    MalwareCheckFactory.create(
-        name="SetupPatternCheck", state=MalwareCheckState.Enabled
-    )
-    check = c.SetupPatternCheck(db_session)
-
-    assert check._extract_setup_contents(tarbuf) is None
-
-
-@pytest.mark.parametrize(
-    ("obj_id", "file_url"), [(None, pretend.stub()), (pretend.stub(), None)]
-)
-def test_scan_missing_kwargs(db_session, obj_id, file_url):
-    MalwareCheckFactory.create(
-        name="SetupPatternCheck", state=MalwareCheckState.Enabled
-    )
-    check = c.SetupPatternCheck(db_session)
-    check.scan(obj_id=obj_id, file_url=file_url)
-
-    assert check._verdicts == []
-
-
-def test_scan_non_sdist(db_session):
-    MalwareCheckFactory.create(
-        name="SetupPatternCheck", state=MalwareCheckState.Enabled
-    )
-    check = c.SetupPatternCheck(db_session)
-
-    file = FileFactory.create(packagetype="bdist_wheel")
-
-    check.scan(obj_id=file.id, file_url=pretend.stub())
-
-    assert check._verdicts == []
-
-
-def test_scan_request_failed(db_session, monkeypatch):
-    requests = pretend.stub(
-        get=pretend.call_recorder(lambda url: pretend.stub(ok=False))
-    )
-    monkeypatch.setattr(c, "requests", requests)
-
-    MalwareCheckFactory.create(
-        name="SetupPatternCheck", state=MalwareCheckState.Enabled
-    )
-    check = c.SetupPatternCheck(db_session)
-
-    file = FileFactory.create(packagetype="sdist")
-
-    check.scan(obj_id=file.id, file_url=pretend.stub())
-
-    assert check._verdicts == []
-
-
-def test_scan_no_setup_contents(db_session, monkeypatch):
-    requests = pretend.stub(
-        get=pretend.call_recorder(
-            lambda url: pretend.stub(ok=True, content=b"not real content")
-        )
-    )
-    monkeypatch.setattr(c, "requests", requests)
-
-    MalwareCheckFactory.create(
-        name="SetupPatternCheck", state=MalwareCheckState.Enabled
-    )
-    check = c.SetupPatternCheck(db_session)
-    monkeypatch.setattr(
-        check, "_extract_setup_contents", pretend.call_recorder(lambda *a: None)
-    )
-
-    file = FileFactory.create(packagetype="sdist")
-
-    check.scan(obj_id=file.id, file_url=pretend.stub())
-
-    assert len(check._verdicts) == 1
-    assert check._verdicts[0].check_id == check.id
-    assert check._verdicts[0].file_id == file.id
-    assert check._verdicts[0].classification == VerdictClassification.Indeterminate
-    assert check._verdicts[0].confidence == VerdictConfidence.High
-    assert (
-        check._verdicts[0].message
-        == "sdist does not contain a suitable setup.py for analysis"
-    )
-
-
-def test_scan_benign_contents(db_session, monkeypatch):
-    requests = pretend.stub(
-        get=pretend.call_recorder(
-            lambda url: pretend.stub(ok=True, content=b"not real content")
-        )
-    )
-    monkeypatch.setattr(c, "requests", requests)
-
-    MalwareCheckFactory.create(
-        name="SetupPatternCheck", state=MalwareCheckState.Enabled
-    )
-    check = c.SetupPatternCheck(db_session)
-    monkeypatch.setattr(
-        check,
-        "_extract_setup_contents",
-        pretend.call_recorder(lambda *a: b"this is a benign string"),
-    )
-
-    file = FileFactory.create(packagetype="sdist")
-
-    check.scan(obj_id=file.id, file_url=pretend.stub())
-
-    assert check._verdicts == []
-
-
-def test_scan_matched_content(db_session, monkeypatch):
-    requests = pretend.stub(
-        get=pretend.call_recorder(
-            lambda url: pretend.stub(ok=True, content=b"not real content")
-        )
-    )
-    monkeypatch.setattr(c, "requests", requests)
-
-    MalwareCheckFactory.create(
-        name="SetupPatternCheck", state=MalwareCheckState.Enabled
-    )
-    check = c.SetupPatternCheck(db_session)
-    monkeypatch.setattr(
-        check,
-        "_extract_setup_contents",
-        pretend.call_recorder(
-            lambda *a: b"this looks suspicious: os.system('cat /etc/passwd')"
-        ),
-    )
-
-    file = FileFactory.create(packagetype="sdist")
-
-    check.scan(obj_id=file.id, file_url=pretend.stub())
-
-    assert len(check._verdicts) == 1
-    assert check._verdicts[0].check_id == check.id
-    assert check._verdicts[0].file_id == file.id
-    assert check._verdicts[0].classification == VerdictClassification.Threat
-    assert check._verdicts[0].confidence == VerdictConfidence.High
-    assert check._verdicts[0].message == "process_spawn_in_setup"
