+++ warehouse/malware/checks/setup_patterns/check.py
@@ -11,16 +11,13 @@
 # limitations under the License.
 
 import os
-import tarfile
-import zipfile
 
-from io import BytesIO
 from textwrap import dedent
 
-import requests
 import yara
 
 from warehouse.malware.checks.base import MalwareCheckBase
+from warehouse.malware.checks.utils import extract_file_content, fetch_url_content
 from warehouse.malware.models import VerdictClassification, VerdictConfidence
 from warehouse.packaging.models import File
 
@@ -48,35 +45,6 @@ class SetupPatternCheck(MalwareCheckBase):
     def _load_yara_rules(self):
         return yara.compile(filepath=self._yara_rule_file)
 
-    def _extract_setup_contents(self, archive_stream):
-        if zipfile.is_zipfile(archive_stream):
-            with zipfile.ZipFile(archive_stream) as zipobj:
-                for name in zipobj.namelist():
-                    parts = os.path.split(name)
-                    if len(parts) == 2 and parts[1] == "setup.py":
-                        return zipobj.read(name)
-                return None
-        else:
-            # NOTE: is_zipfile doesn't rewind the fileobj it's given.
-            archive_stream.seek(0)
-
-            # NOTE: We don't need to perform a sanity check on
-            # the (presumed) tarfile's compression here, since we're
-            # extracting from a stream that's already gone through
-            # upload validation.
-            # See _is_valid_dist_file in forklift/legacy.py.
-            try:
-                with tarfile.open(fileobj=archive_stream) as tarobj:
-                    member = tarobj.next()
-                    while member:
-                        parts = os.path.split(member.name)
-                        if len(parts) == 2 and parts[1] == "setup.py":
-                            return tarobj.extractfile(member).read()
-                        member = tarobj.next()
-                    return None
-            except tarfile.TarError:
-                return None
-
     def scan(self, **kwargs):
         file_id = kwargs.get("obj_id")
         file_url = kwargs.get("file_url")
@@ -92,15 +60,8 @@ class SetupPatternCheck(MalwareCheckBase):
             # we have nothing to perform.
             return
 
-        # TODO(ww): Handle connection errors, timeouts?
-        # The check harness will catch everything eventually,
-        # so these might not be worth handling explicitly.
-        file_request = requests.get(file_url)
-        if not file_request.ok:
-            # TODO: Indeterminate verdict here?
-            return
-
-        setup_py_contents = self._extract_setup_contents(BytesIO(file_request.content))
+        archive_stream = fetch_url_content(file_url)
+        setup_py_contents = extract_file_content(archive_stream, "setup.py")
         if setup_py_contents is None:
             self.add_verdict(
                 file_id=file_id,
